backend/requirements.txt
altair==5.5.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
distro==1.9.0
fastapi==0.120.4
filelock==3.20.0
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lxml==6.0.2
MarkupSafe==3.0.3
mpmath==1.3.0
narwhals==2.10.2
networkx==3.5
numpy==2.3.4
openai==2.7.1
packaging==25.0
pandas==2.3.3
pgvector==0.4.1
pillow==12.0.0
protobuf==6.33.0
psycopg2-binary==2.9.11
pyarrow==21.0.0
pydantic==2.12.3
pydantic_core==2.41.4
pydeck==0.9.1
PyMuPDF==1.26.6
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2025.10.23
requests==2.32.5
rpds-py==0.28.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.3
sentence-transformers==5.1.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.49.3
streamlit==1.51.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
toml==0.10.2
torch==2.9.0
tornado==6.5.2
tqdm==4.67.1
transformers==4.57.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
